backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<26.0
django-ratelimit>=4.1,<5.0
django-csp>=3.8,<4.0
psycopg[binary]>=3.2,<4.0
argon2-cffi>=23.1,<26.0
bleach>=6.1,<7.0
stripe>=10.0,<13.0
gunicorn>=22.0,<24.0
